Create initial project strucutre and move files around.
