--- Rambler - Text_adventure_game/Project_task_list.txt
@@ -1,38 +1,0 @@
-The basis of the game is a player character that handles events by making decisions untill the HP is 0. 
-It is like an RPG with a character with HP, EXP, etc. 
-Events hold the main content of the game and is presented to the player in random order. therefor it is non-linear. 
-Events will be text-pieces that provide in-game lore and make up the main story of the game. 
-----------------------------------------------------------------------------------------------
-| The first task will be to create a framework for creation, maintenance, viewing, and deleting events. 
-| We will need a framework for easy handling of events so that it will be possible from a development perspective expand the story of the game. 
-| There will be uses cases + test cases for all 4 (3) actions of the framework. 
-| !!!!
-----------------------------------------------------------------------------------------------
-| UC 1: Event framework: Create and save.
-| Events must be persisted as .txt files in a directory that is part of the game files.
-| For now the directory will be a subdirectory under the main project directory. 
-| the directory should be named "c:/...../game_events". 
-| An event should be given a descriptive name and suffixed by a 4 digit counter that functions as an ID-number eg: "first_event_0001.txt"
-| there might be an concurrency issue in the future. It might be worth exploring using a Datebase. For now it will not change.
------------------------------------------------------------------------------------------------
-| UC 2: Event framework: View and Maintain.
-| It must be possible to view the contents of an event. It must be possible to edit an event. The view and edit "modes" should be switchable via a button.
-| It must not be possible to edit "key" fields such as id-numbers. 
-| The body text of the event must be viweved in a text field. 
-| In edit mode it must be possible to save the changes to the event. 
------------------------------------------------------------------------------------------------
-| UC 3: Event framework: Delete an event
-| It must be possible to delete an event.
-| this should be done via the view and maintan dialog. 
-| the user must be prompted with an "are you sure?" pop-up when trying to delete an event. 
------------------------------------------------------------------------------------------------
------------------------------------------------------------------------------------------------
-| UC 1.1: ID number 
-| The ID-number of the events must be non-reuseable, unique number. 
-| It must use the number range 0000-9999. It should be possible to expand the number range after 9999 to A001 etc. 
------------------------------------------------------------------------------------------------
-| UC 1.2: Events
-| An event should hold relevant header data: name, id, related events, game version, view-order. 
-| The main body of an event should consist of a text-string, with a configurable minimum and maximum length. 
-| the main body of the text should only accept ascii charathers? 
-
