+++ test/get_next_number_test.py
@@ -1,0 +1,9 @@
+import unittest
+import get_next_number
+import os, os.path
+
+#! TODO: finish this unit test when event persistence is done.
+class Get_nxt_number_test (unittest.TestCase):
+    def set_up (self):
+        print("setting up test")
+
